+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - gui: Fixed decimal input on ios [https://github.com/toniebox-reverse-engineering/teddycloud_web/issues/289](https://github.com/toniebox-reverse-engineering/teddycloud_web/issues/289)
 - gui: TeddyStudio: Added Image Bottom/Left margin [https://github.com/toniebox-reverse-engineering/teddycloud_web/issues/284](https://github.com/toniebox-reverse-engineering/teddycloud_web/issues/284)
 - gui: TeddyStudio: Added font selection for label text [https://github.com/toniebox-reverse-engineering/teddycloud_web/pull/299](https://github.com/toniebox-reverse-engineering/teddycloud_web/pull/299)
+- gui: TeddyStudio: Added bulk-add tonies modal [https://github.com/toniebox-reverse-engineering/teddycloud_web/pull/300](https://github.com/toniebox-reverse-engineering/teddycloud_web/pull/300)
 - gui: added Custom-Model-Editor to create and edit custom models
 - gui: added custom image management to library
 - gui: added selection for custom and original images for TeddyStudio and Custom-Model-Editor